Add dockerfile for replication.py function
Creates a dockerfile utilizing python version 3.0 to run the
replication.py script. Also adds two functions for counting
number of times a pattern exist in a genomic text partition and
a dict keeping track of the number of each times a specific
pattern exists to find the most occuring k-mer pattern.